LayoutTests/platform/wpe/fast/attachment/attachment-dom-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: HTMLAttachmentElement
This tests that attachments have an instance type of HTMLElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
